Fix circular dep
