+++ src/clientSideScene/ClientSideSceneComp.tsx
@@ -27,7 +27,8 @@ import { getConstraintInfoKw } from '@src/lang/std/sketch'
 import type { ConstrainInfo } from '@src/lang/std/stdTypes'
 import { topLevelRange } from '@src/lang/util'
 import type { CallExpressionKw, Expr, PathToNode } from '@src/lang/wasm'
-import { defaultSourceRange, parse, recast, resultIsOk } from '@src/lang/wasm'
+import { parse, recast, resultIsOk } from '@src/lang/wasm'
+import { defaultSourceRange } from '@src/lang/sourceRange'
 import { cameraMouseDragGuards } from '@src/lib/cameraControls'
 import {
   codeManager,
