+++ src/components/AstExplorer.tsx
@@ -5,7 +5,7 @@ import { getNodeFromPath } from '@src/lang/queryAst'
 import { getNodePathFromSourceRange } from '@src/lang/queryAstNodePathUtils'
 import { codeRefFromRange } from '@src/lang/std/artifactGraph'
 import { topLevelRange } from '@src/lang/util'
-import { defaultSourceRange } from '@src/lang/wasm'
+import { defaultSourceRange } from '@src/lang/sourceRange'
 import { codeToIdSelections } from '@src/lib/selections'
 import { editorManager, kclManager } from '@src/lib/singletons'
 import { trap } from '@src/lib/trap'
